# Optimize hooks, reduce library size and get rid of edge case errors and memory leaks