--- a/src/hooks/useEventCallback.ts
+++ b/src/hooks/useEventCallback.ts
@@ -1,14 +1,12 @@
-import { useRef, useEffect, useCallback } from "react";
+import { useRef } from "react";
 
 // Saves incoming handler to the ref in order to avoid "useCallback hell"
-function useEventCallback<T>(handler?: (value: T) => void): (value: T) => void {
+export function useEventCallback<T>(handler?: (value: T) => void): (value: T) => void {
   const callbackRef = useRef(handler);
-
-  useEffect(() => {
-    callbackRef.current = handler;
+  const fn = useRef((value: T) => {
+    callbackRef.current && callbackRef.current(value);
   });
+  callbackRef.current = handler;
 
-  return useCallback((value: T) => callbackRef.current && callbackRef.current(value), []);
+  return fn.current;
 }
-
-export { useEventCallback };