This diff makes the runtime faster and the library size smaller

|Before|After|
|---|---|
| ![image](https://user-images.githubusercontent.com/1178825/127752345-e76fc4b1-780e-4746-9a81-3b3c1e885f5c.png) | ![image](https://user-images.githubusercontent.com/1178825/128600170-5d286787-af47-409f-ad9d-997ef59a1b1d.png) |
| ![image](https://user-images.githubusercontent.com/1178825/127752357-39d94d53-7395-4d91-a361-80a788f72e4c.png) | ![image](https://user-images.githubusercontent.com/1178825/128600184-e9b3b1a7-504e-4b7e-be8b-b9c5dab09331.png) |

As we get rid of `useEffect` calls, it saves time in runtime. 

It also prevents us from the memory leak and potential errors


The behaviour isn't changed:

https://user-images.githubusercontent.com/1178825/128600229-9f30f0ae-5cf6-4c31-ae57-62cb17e3bdce.mov