@xnimorz 2021-08-07
I'm not sure how to open the multitouch PR as it's based on the commit from this PR (you can see it here atm https://github.com/xnimorz/react-colorful/pull/1/files, I'll publish it to your repo after this diff)