@xnimorz 2021-08-30
Updated data: 
![image](https://user-images.githubusercontent.com/1178825/131318643-2d2368ba-5679-41e2-a879-8daaca889525.png)